+++ CHANGELOG.md
@@ -10,11 +10,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `Dockerfile`
 - Add `tsconfig.json`. Move typescript config from `gulpfile.js`.
 - Add PHP `7.3` to travis-ci build matrix
+- Add jQuery version 3 (with Migrate plugin)
 
 ### Changed
 - Replace hard-coded domain path in core translation setup with one grabbed from `style.css`
 - Upgrade gulp to version 4
 - Update scripts to use typescript namespaces and classes
+
+### Removed
+- Remove jQuery version shipped with WordPress core
 
 ## 0.9.2 - 2018-10-08